Fix note type not matching the selected type error
Problem:
The new card doesn't always use the selected note type

Issue:
The code will set the global note type but not set the note type for the current deck. When creating a new note, the note type for the current deck is used.

Fix:
When creating a new note, use the current global note type.